+++ conftest.py
@@ -89,7 +89,7 @@ def mock_cisco_api_authentication_server(monkeypatch):
     monkeypatch.setattr(
         requests,
         "post",
-        lambda url, params=None: mock_post_response()
+        lambda url, params=None, proxies=None: mock_post_response()
     )