+++ django_project/celery.py
@@ -13,7 +13,7 @@ from raven.contrib.celery import register_signal, register_logger_signal
 class Celery(celery.Celery):
 
     def on_configure(self):
-        if settings.PDB_ENABLE_SENTRY:  # ignore for coverage
+        if settings.PDB_SENTRY_DSN:  # ignore for coverage
             client = raven.Client(settings.PDB_SENTRY_DSN)
             client.release = raven.fetch_git_sha(os.path.dirname(os.pardir))